Mission Planner 1.2.46 add binary log support add new self describing log format disable dtr when using px4 fix failsafe screen fix AC accel calib mag calib is no 60 seconds from 30 firmware > add latest fw remember last lat,lng,alt for main map. fix swarm initial location

+++ CurrentState.cs
@@ -33,9 +33,9 @@ namespace ArdupilotMega
 
         // position
         [DisplayText("Latitude (dd)")]
-        public float lat { get; set; }
+        public double lat { get; set; }
         [DisplayText("Longitude (dd)")]
-        public float lng { get; set; }
+        public double lng { get; set; }
         [DisplayText("Altitude (dist)")]
         public float alt { get { return (_alt - altoffsethome) * multiplierdist; } set { _alt = value; } }
         DateTime lastalt = DateTime.MinValue;
